Makes better use of json-schema to test the changes to the record resource
Also mark the YAML test for the record resource as xfail for now. This is
because this test currently fails for registers that use "integers" for keys.
This happens because we do not surround our keys with quotes in our API and the python yaml
library attempts to convert this into various datatypes. We then get a string comparison failure.
We need to work out if this is a problem with our API or the YAML library before removing the xfail.