Don't enforce the current URL versioning structure
We expect that v1 APIs will be served at /v1/ and v2 at /v2/ etc.
But there may be cases where this is not followed, and it should
still be possible to test for conformance in other ways.

For example, for GOV.UK registers we intend to serve in-progress
implementation of the upcoming API version at (/next). In this case
we should be able to run the conformance tests against that implementation,
without forcing the same approach and release timelines on other
implementations of the specification.